Fold hardhat package into SDK package Reorganize SDK consts and utils into separate folders
