--- typescript/hardhat/.gitignore
@@ -1,2 +1,0 @@
-dist/
-cache/
