--- typescript/hardhat/.prettierrc
@@ -1,5 +1,0 @@
-{
-  "tabWidth": 2,
-  "singleQuote": true,
-  "trailingComma": "all"
-}
