--- typescript/hardhat/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "@abacus-network/hardhat",
-  "description": "Hardhat related tools for the Abacus network",
-  "version": "0.2.3",
-  "dependencies": {
-    "@abacus-network/core": "^0.2.3",
-    "@abacus-network/deploy": "^0.2.3",
-    "@abacus-network/sdk": "^0.2.3",
-    "@abacus-network/utils": "^0.2.3",
-    "@nomiclabs/hardhat-ethers": "^2.0.5",
-    "@nomiclabs/hardhat-waffle": "^2.0.2",
-    "ethereum-waffle": "^3.2.2",
-    "ethers": "^5.6.8",
-    "hardhat": "^2.8.4"
-  },
-  "devDependencies": {
-    "prettier": "^2.4.1",
-    "ts-node": "^10.8.0",
-    "typescript": "^4.7.2"
-  },
-  "directories": {
-    "src": "src",
-    "test": "test"
-  },
-  "homepage": "https://www.useabacus.network",
-  "keywords": [
-    "Abacus",
-    "Typescript",
-    "Hardhat"
-  ],
-  "license": "Apache-2.0",
-  "main": "dist/index.js",
-  "prepublish": "yarn build",
-  "repository": "https://github.com/abacus-network/abacus-monorepo",
-  "scripts": {
-    "build": "tsc",
-    "check": "tsc --noEmit",
-    "prettier": "prettier --write ./src ./test",
-    "test": "hardhat test"
-  },
-  "types": "dist/index.d.ts",
-  "files": [
-    "/dist"
-  ]
-}
