+++ typescript/sdk/package.json
@@ -6,23 +6,32 @@
     "@abacus-network/celo-ethers-provider": "^0.1.0",
     "@abacus-network/core": "^0.2.3",
     "@abacus-network/utils": "^0.2.3",
+    "@types/debug": "^4.1.7",
+    "debug": "^4.3.4",
     "ethers": "^5.6.8"
   },
   "devDependencies": {
+    "@nomiclabs/hardhat-ethers": "^2.0.5",
+    "@nomiclabs/hardhat-waffle": "^2.0.2",
     "@types/node": "^16.9.1",
     "chai": "^4.3.6",
     "dotenv": "^10.0.0",
+    "ethereum-waffle": "^3.2.2",
     "fs": "0.0.1-security",
+    "hardhat": "^2.8.4",
     "mocha": "^9.2.2",
     "prettier": "^2.4.1",
     "sinon": "^13.0.2",
     "typescript": "^4.7.2"
   },
+  "files": [
+    "/dist"
+  ],
   "homepage": "https://www.useabacus.network",
   "keywords": [
     "Abacus",
-    "Typescript",
-    "SDK"
+    "SDK",
+    "Typescript"
   ],
   "license": "Apache-2.0",
   "main": "dist/index.js",
@@ -30,12 +39,11 @@
   "scripts": {
     "build": "tsc",
     "check": "tsc --noEmit",
+    "clean": "rm -rf ./dist",
     "prepublishOnly": "yarn build",
     "prettier": "prettier --write ./src ./test",
-    "test": "mocha --config .mocharc.json './test/**/*.test.ts'"
+    "test": "mocha --config .mocharc.json './test/**/*.test.ts'",
+    "test:hardhat": "hardhat test"
   },
-  "types": "dist/index.d.ts",
-  "files": [
-    "/dist"
-  ]
+  "types": "dist/index.d.ts"
 }
