+++ typescript/sdk/src/consts/chains.ts
@@ -1,0 +1,26 @@
+/**
+ * Enumeration of Abacus supported chains
+ */
+export enum Chains { // must be string type to be used with Object.keys
+  arbitrum = 'arbitrum',
+  alfajores = 'alfajores',
+  bsc = 'bsc',
+  mumbai = 'mumbai',
+  kovan = 'kovan',
+  goerli = 'goerli',
+  fuji = 'fuji',
+  celo = 'celo',
+  ethereum = 'ethereum',
+  avalanche = 'avalanche',
+  optimism = 'optimism',
+  polygon = 'polygon',
+  bsctestnet = 'bsctestnet',
+  arbitrumrinkeby = 'arbitrumrinkeby',
+  optimismkovan = 'optimismkovan',
+  auroratestnet = 'auroratestnet',
+  test1 = 'test1',
+  test2 = 'test2',
+  test3 = 'test3',
+}
+
+export const AllChains = Object.keys(Chains) as Array<keyof typeof Chains>;
