+++ typescript/sdk/src/consts/environments/mainnet.json
