+++ typescript/sdk/src/consts/environments/test.json
