+++ typescript/sdk/src/consts/environments/testnet.json
