+++ typescript/sdk/src/consts/environments/testnet2.json
