--- typescript/sdk/src/ethers/index.ts
@@ -1,1 +1,0 @@
-export { RetryProvider, RetryJsonRpcProvider } from './retry-provider';
