+++ typescript/sdk/src/hardhat/index.ts
@@ -1,0 +1,2 @@
+export * from './TestCoreApp';
+export * from './TestCoreDeploy';
