+++ typescript/sdk/src/index.ts
@@ -1,6 +1,19 @@
-export { AbacusApp } from './app';
-export { chainMetadata } from './chain-metadata';
-export { addSignerToConnection, chainConnectionConfigs } from './chains';
+export { AllChains, Chains } from './consts/chains';
+export { chainMetadata } from './consts/chainMetadata';
+export { chainConnectionConfigs } from './consts/chainConnectionConfigs';
+export {
+  ChainMap,
+  ChainName,
+  CompleteChainMap,
+  Connection,
+  NameOrDomain,
+  RemoteChainMap,
+  Remotes,
+  TestChainNames,
+} from './types';
+export { ChainNameToDomainId, DomainIdToChainName } from './domains';
+
+export { AbacusApp } from './AbacusApp';
 export {
   AbacusAddresses,
   AbacusContracts,
@@ -27,33 +40,36 @@ export {
   resolveId,
   resolveNetworks,
 } from './core';
-export { RetryJsonRpcProvider, RetryProvider } from './ethers';
 export {
   Annotated,
   getEvents,
   queryAnnotatedEvents,
   TSContract,
 } from './events';
+export { ChainConnection, IChainConnection, MultiProvider } from './provider';
+export { BeaconProxyAddresses, ProxiedContract, ProxyAddresses } from './proxy';
+export { Router, RouterContracts, RouterFactories } from './router';
+
 export {
   DefaultTokenPriceGetter,
   InterchainGasCalculator,
   TokenPriceGetter,
 } from './gas';
-export { ChainConnection, IChainConnection, MultiProvider } from './provider';
-export { BeaconProxyAddresses, ProxiedContract, ProxyAddresses } from './proxy';
-export { Router, RouterContracts, RouterFactories } from './router';
+
 export {
-  AllChains,
-  ChainMap,
-  ChainName,
-  ChainNameToDomainId,
-  Chains,
-  CompleteChainMap,
-  Connection,
-  DomainIdToChainName,
-  NameOrDomain,
-  RemoteChainMap,
-  Remotes,
-  TestChainNames,
-} from './types';
-export { objMap, objMapEntries, promiseObjAll, utils } from './utils';
+  objMap,
+  objMapEntries,
+  promiseObjAll,
+  addSignerToConnection,
+  RetryJsonRpcProvider,
+  RetryProvider,
+} from './utils';
+export * as utils from './utils';
+
+export {
+  TestCoreApp,
+  TestCoreContracts,
+  TestInboxContracts,
+  TestOutboxContracts,
+  TestCoreDeploy,
+} from './hardhat';
