+++ typescript/sdk/src/types.ts
@@ -1,52 +1,12 @@
-import { ethers } from 'ethers';
+import type { ethers } from 'ethers';
 
-import { chainMetadata } from './chain-metadata';
+import type { Chains } from './consts/chains';
 
-/**
- * RPC Pagination information for Polygon
- */
-export interface Pagination {
-  blocks: number;
-  from: number;
-}
-
-/**
- * Enumeration of Abacus supported chains
- */
-export enum Chains { // must be string type to be used with Object.keys
-  arbitrum = 'arbitrum',
-  alfajores = 'alfajores',
-  bsc = 'bsc',
-  mumbai = 'mumbai',
-  kovan = 'kovan',
-  goerli = 'goerli',
-  fuji = 'fuji',
-  celo = 'celo',
-  ethereum = 'ethereum',
-  avalanche = 'avalanche',
-  optimism = 'optimism',
-  polygon = 'polygon',
-  bsctestnet = 'bsctestnet',
-  arbitrumrinkeby = 'arbitrumrinkeby',
-  optimismkovan = 'optimismkovan',
-  auroratestnet = 'auroratestnet',
-  test1 = 'test1',
-  test2 = 'test2',
-  test3 = 'test3',
-}
 export type ChainName = keyof typeof Chains;
 export type CompleteChainMap<Value> = Record<ChainName, Value>;
 export type ChainMap<Chain extends ChainName, Value> = Record<Chain, Value>;
-
 export type TestChainNames = 'test1' | 'test2' | 'test3';
 
-export const AllChains = Object.keys(Chains) as ChainName[];
-export const DomainIdToChainName = Object.fromEntries(
-  AllChains.map((chain) => [chainMetadata[chain].id, chain]),
-);
-export const ChainNameToDomainId = Object.fromEntries(
-  AllChains.map((chain) => [chain, chainMetadata[chain].id]),
-) as CompleteChainMap<number>;
 export type NameOrDomain = ChainName | number;
 
 export type Remotes<
@@ -59,15 +19,5 @@ export type RemoteChainMap<
   LocalChain extends Chain,
   Value,
 > = Record<Remotes<Chain, LocalChain>, Value>;
-
-/**
- * A Domain (and its characteristics)
- */
-export type ChainMetadata = {
-  id: number;
-  finalityBlocks: number;
-  nativeTokenDecimals?: number;
-  paginate?: Pagination;
-};
 
 export type Connection = ethers.providers.Provider | ethers.Signer;
