+++ typescript/sdk/src/utils/.eslintrc
@@ -1,0 +1,5 @@
+{
+  "rules": {
+    "@typescript-eslint/explicit-module-boundary-types": ["off"]
+  }
+}
