+++ typescript/sdk/src/utils/RetryProvider.ts
@@ -2,22 +2,21 @@
 //
 // Mostly taken from the removed version that was in ethers.js
 // See: https://github.com/ethers-io/ethers.js/discussions/3006
-import { BaseProvider, JsonRpcProvider } from '@ethersproject/providers';
 import { ethers } from 'ethers';
 
-import { retryAsync } from '@abacus-network/utils/dist/src/utils';
+import { utils } from '@abacus-network/utils';
 
 export type RetryOptions = {
   // The wait interval in between
   interval: number;
 
-  // Maximum number of times to rety
+  // Maximum number of times to retry
   retryLimit: number;
 };
 
-export class RetryProvider extends BaseProvider {
+export class RetryProvider extends ethers.providers.BaseProvider {
   constructor(
-    readonly provider: BaseProvider,
+    readonly provider: ethers.providers.BaseProvider,
     readonly retryOptions: RetryOptions,
   ) {
     super(provider.getNetwork());
@@ -27,7 +26,7 @@ export class RetryProvider extends BaseProvider {
 
   // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
   perform(method: string, params: any): Promise<any> {
-    return retryAsync(
+    return utils.retryAsync(
       () => this.provider.perform(method, params),
       this.retryOptions.retryLimit,
       this.retryOptions.interval,
@@ -36,16 +35,16 @@ export class RetryProvider extends BaseProvider {
 }
 
 // Need this separate class for JsonRpcProvider to still expose `getSigner`, so will retry at the request level
-export class RetryJsonRpcProvider extends JsonRpcProvider {
+export class RetryJsonRpcProvider extends ethers.providers.JsonRpcProvider {
   constructor(
-    readonly provider: JsonRpcProvider,
+    readonly provider: ethers.providers.JsonRpcProvider,
     readonly retryOptions: RetryOptions,
   ) {
     super(provider.connection, provider.network);
   }
 
   async send(method: string, params: Array<any>): Promise<any> {
-    return retryAsync(
+    return utils.retryAsync(
       async () => this.provider.send(method, params),
       this.retryOptions.retryLimit,
       this.retryOptions.interval,
