+++ typescript/sdk/src/utils/objects.ts
@@ -1,0 +1,29 @@
+// TODO move to utils package
+
+export function objMapEntries<K extends string, I = any, O = any>(
+  obj: Record<K, I>,
+  func: (k: K, _: I) => O,
+): [K, O][] {
+  return Object.entries<I>(obj).map(([k, v]) => [k as K, func(k as K, v)]);
+}
+
+// Map over the values of the object
+export function objMap<K extends string, I = any, O = any>(
+  obj: Record<K, I>,
+  func: (k: K, _: I) => O,
+) {
+  return Object.fromEntries<O>(objMapEntries<K, I, O>(obj, func)) as Record<
+    K,
+    O
+  >;
+}
+
+// promiseObjectAll :: {k: Promise a} -> Promise {k: a}
+export const promiseObjAll = <K extends string, V>(object: {
+  [key in K]: Promise<V>;
+}): Promise<Record<K, V>> => {
+  const promiseList = Object.entries(object).map(([name, promise]) =>
+    (promise as Promise<V>).then((result) => [name, result]),
+  );
+  return Promise.all(promiseList).then(Object.fromEntries);
+};
