+++ typescript/sdk/src/utils/time.ts
@@ -1,0 +1,11 @@
+// TODO move to utils package
+
+/**
+ * Sleep async for some time.
+ *
+ * @param ms the number of milliseconds to sleep
+ * @returns A delay promise
+ */
+export function delay(ms: number): Promise<void> {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
