+++ typescript/sdk/tsconfig.json
@@ -4,6 +4,5 @@
     "outDir": "./dist/",
     "rootDir": "./src/"
   },
-  "exclude": ["./node_modules/", "./dist/", "./test/"],
-  "include": ["./src/*.ts", "./src/**/*.ts"]
+  "include": ["./src/**/*.ts"]
 }
